BREAKING: update to fkeycapture 1.3.0 + smaller config fixes

+++ package/upload.sh
@@ -1,4 +1,8 @@
 source .env
+if [ -z ${PYPI_TOKEN+x} ]; then
+	echo 'Cannot upload without a pypi token. Please set one in a .env file next to this script with the name of `PYPI_TOKEN`.'
+	exit 1
+fi
 echo "🧽 Pruging old package files... 🧽"
 rm -rf dist
 mkdir dist
